requirements.txt
# Core dependencies
csfloat-api>=1.3.0
aiohttp>=3.9.0
aiohttp-socks>=0.8.0

# Web framework
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6
jinja2>=3.1.2

# Database & ORM
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2

# HTTP client for advanced orders
httpx>=0.26.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0  # Pin to 4.x due to passlib compatibility

# WebSocket support
websockets>=12.0
